shims/linux: move epoll and eventfd into their own files, together with their FD types
